changed double to float

+++ include/casting.h
@@ -6,7 +6,7 @@
 /*   By: gcrisp <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/21 13:26:47 by gcrisp            #+#    #+#             */
-/*   Updated: 2025/02/21 13:27:40 by gcrisp           ###   ########.fr       */
+/*   Updated: 2025/02/21 15:33:00 by gcrisp           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,8 +18,8 @@
 
 typedef struct s_point
 {
-	double	x;
-	double	y;
+	float	x;
+	float	y;
 }	t_point;
 
 typedef struct s_boundary
@@ -34,14 +34,14 @@ typedef struct s_ray
 	t_point	*dir;
 }	t_ray;
 
-t_boundary	*new_boundary(double x1, double y1, double x2, double y2);
+t_boundary	*new_boundary(float x1, float y1, float x2, float y2);
 void		free_boundary(t_boundary *bound);
 void		free_boundary_array(t_boundary *bounds, size_t num_bounds);
 
-t_ray		*new_ray(double x, double y, double degs);
+t_ray		*new_ray(float x, float y, float degs);
 void		free_ray(t_ray *ray);
 
-t_point		*new_point(double x, double y);
+t_point		*new_point(float x, float y);
 
 int			has_intersection(t_boundary *bound, t_ray *ray);
 t_point		*get_intersection(t_boundary *bound, t_ray *ray);
