+++ srcs/map.c
@@ -6,13 +6,13 @@
 /*   By: gcrisp <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/21 12:27:00 by gcrisp            #+#    #+#             */
-/*   Updated: 2025/02/21 14:16:00 by gcrisp           ###   ########.fr       */
+/*   Updated: 2025/02/21 15:37:38 by gcrisp           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "map.h"
 
-t_map	*new_map(double p_x, double p_y, t_boundary *bounds, size_t num_bounds)
+t_map	*new_map(float p_x, float p_y, t_boundary *bounds, size_t num_bounds)
 {
 	t_map	*out;
